LayoutLMv3-project/requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision

# Lightning for training structure
pytorch-lightning>=2.0.0

# Transformers & Document AI
transformers>=4.28.0
huggingface-hub

# Image Processing & Augmentation
Pillow
opencv-python
albumentations>=1.3.0
numpy

# Evaluation Metrics for NER/Token Classification
seqeval

# Hugging Face ecosystem (Essential for Donut)
transformers>=4.25.1
datasets
accelerate

# Utilities
tqdm
matplotlib
sentencepiece

# pip install -r requirements.txt
# pip install 'git+https://github.com/facebookresearch/detectron2.git'
